feat: add record detail components and functionality
- Introduced RecordDetailHeader, RecordDetailMainColumn, RecordDetailSidebar, and SimilarRunCard components for displaying detailed run information.
- Implemented state management for comments, likes, shares, and video playback synchronization in RecordDetailPage.
- Added helper functions for managing party loadouts and YouTube video commands.
- Created UI components for displaying character avatars, loadout entries, and similarity reasons.
- Established configuration constants for styling and layout.
- Defined types for party loadout entries and similar action states.
